+++ tests/cli/test_cli_query.py
@@ -18,54 +18,46 @@ def nqueries(request):
 
 	Value is derived from argument to the "testdb_nqueries" marker, if any. This can be set on
 	specific test functions to improve speed by only using a subset of the query files.
+	(Currently not overridden anywhere).
 
 	Based on this example:
 	https://docs.pytest.org/en/6.2.x/fixture.html#using-markers-to-pass-data-to-fixtures
+
+	Note than with slice notation, `[:None]` is the same as `[:]`.
 	"""
 	marker = request.node.get_closest_marker("testdb_nqueries")
 	return None if marker is None else marker.args[0]
 
 @pytest.fixture()
-def query_files(testdb_queries, nqueries):
+def query_files(testdb_query_files, nqueries):
 	"""Paths to query files."""
-	files = [query['file'] for query in testdb_queries]
-	return files if nqueries is None else files[:nqueries]
+	return testdb_query_files[:nqueries]
 
-@pytest.fixture()
-def results(testdb_results, nqueries):
-	"""Results object to compare output to."""
-	if nqueries is None:
-		return testdb_results
+@pytest.fixture(name='make_args')
+def make_args_factory(testdb_files):
 
-	results = copy(testdb_results)
-	results.items = results.items[:nqueries]
-	return results
+	def make_args(query_files=None, sigfile=None, output=None, outfmt=None, strict=False):
+		"""Make command line arguments for query file."""
 
+		args = [f'--db={testdb_files["root"]}', 'query']
+		args.append('--strict' if strict else '--no-strict')
 
-def make_args(query_files=None, sigfile=None, db=None, output=None, outfmt=None, strict=False):
-	"""Make command line arguments for query file."""
-	args = []
+		if output is not None:
+			args.append(f'--output={output}')
+		if outfmt is not None:
+			args.append(f'--outfmt={outfmt}')
+		if sigfile is not None:
+			args.append(f'--sigfile={sigfile}')
 
-	if db is not None:
-		args.append(f'--db={db}')
+		if query_files is not None:
+			args.extend([str(f.path) for f in query_files])
 
-	args.append('query')
-	args.append('--strict' if strict else '--no-strict')
+		return args
 
-	if output is not None:
-		args.append(f'--output={output}')
-	if outfmt is not None:
-		args.append(f'--outfmt={outfmt}')
-	if sigfile is not None:
-		args.append(f'--sigfile={sigfile}')
-
-	if query_files is not None:
-		args.extend([str(f.path) for f in query_files])
-
-	return args
+	return make_args
 
 
-def check_results(results_file, out_fmt,ref_results):
+def check_results(results_file, out_fmt, ref_results):
 	"""Check results output matches reference QueryResults object."""
 	if out_fmt == 'json':
 		with open(results_file) as fh:
@@ -77,45 +69,64 @@ def check_results(results_file, out_fmt,ref_results):
 		raise ValueError(f'Invalid out_fmt {out_fmt!r}')
 
 
-@pytest.mark.parametrize('out_fmt', ['csv', 'json'])
-def test_full_query(testdb_files, results, query_files, out_fmt, tmp_path):
+# @pytest.mark.parametrize('out_fmt', ['csv', 'json'])
+# @pytest.mark.parametrize('testdb_queries_gzipped', [False, True], indirect=True)
+@pytest.mark.parametrize(
+	['out_fmt', 'testdb_results', 'testdb_queries_gzipped'],
+	[
+		('json', 'non_strict', False),
+		('csv', 'non_strict', False),
+		('json', 'strict', False),
+		('csv', 'strict', False),
+		('json', 'non_strict', True),
+	],
+	indirect=['testdb_results', 'testdb_queries_gzipped'],
+)
+def test_full_query(make_args, testdb_results, nqueries, query_files, out_fmt, tmp_path):
 	"""Run a full query using the command line interface."""
+
+	# Modify results object to match file names and possibly reduced # of queries
+	ref_results = copy(testdb_results)
+	ref_results.items = ref_results.items[:nqueries]
+	for item, file in zip_strict(ref_results.items, query_files):
+		item.input = QueryInput.convert(file)
 
 	results_file = tmp_path / ('results.' + out_fmt)
 
 	args = make_args(
 		query_files=query_files,
-		db=testdb_files['root'],
 		output=results_file,
 		outfmt=out_fmt,
-		strict=results.params.classify_strict,
+		strict=ref_results.params.classify_strict,
 	)
 
 	result = invoke_cli(args)
 	assert result.exit_code == 0
 
-	check_results(results_file, out_fmt, results)
+	check_results(results_file, out_fmt, ref_results)
 
 
-@pytest.mark.parametrize('out_fmt', ['csv', 'json'])
-def test_sigfile(testdb_files, testdb_query_signatures, results, out_fmt, tmp_path):
+# Not really necessary to check all combinations of parameters.
+@pytest.mark.parametrize('out_fmt', ['json'])
+@pytest.mark.parametrize('testdb_results', ['non_strict'], indirect=True)
+def test_sigfile(make_args, testdb_files, testdb_query_signatures, testdb_results, out_fmt, tmp_path):
 	"""Test using signature file instead of parsing genome files."""
 
 	results_file = tmp_path / ('results.' + out_fmt)
 
 	# Modify results object inputs to match signature file
-	for item, id_ in zip_strict(results.items, testdb_query_signatures.ids):
+	ref_results = copy(testdb_results)
+	for item, id_ in zip_strict(ref_results.items, testdb_query_signatures.ids):
 		item.input = QueryInput(id_)
 
 	args = make_args(
 		sigfile=testdb_files['query_signatures'],
-		db=testdb_files['root'],
 		output=results_file,
 		outfmt=out_fmt,
-		strict=results.params.classify_strict,
+		strict=ref_results.params.classify_strict,
 	)
 
 	result = invoke_cli(args)
 	assert result.exit_code == 0
 
-	check_results(results_file, out_fmt, results)
+	check_results(results_file, out_fmt, ref_results)